fix crash when deleting new bookmarks
(cherry picked from commit 70a2e0c31f8b0420ba4cbb94a4118e59f2d49c73)

+++ app/src/main/java/com/github/andreyasadchy/xtra/repository/ApiRepository.kt
@@ -498,7 +498,7 @@ class ApiRepository @Inject constructor(
                                 addProperty("channel_id", channelId)
                                 addProperty("broadcast_id", streamId)
                                 addProperty("player", "site")
-                                addProperty("user_id", userId?.toInt())
+                                addProperty("user_id", userId?.toLong())
                             })
                         }
                         val spadeRequest = Base64.encodeToString(json.toString().toByteArray(), Base64.NO_WRAP)
